[analyzer] Factor out NoOwnershipChangeVisitor
In preparation for adding essentially the same visitor to StreamChecker,
this patch factors this visitor out to a common header.

Its true that for the time being NoOwnershipChangeVisitor and
NoStateChangeVisitor play a very similar role, so it'd be reasonable
desire to merge these. My argument against that is that if we check for
a property change that is not a resource this distinction could still be
useful.
